add actix

+++ src/models/response_model/user_response.rs
@@ -5,7 +5,7 @@ use super::common_models::{Pagination, ResponseMetadata};
 #[derive(Serialize, Deserialize)]
 pub struct UserResponse {
     pub results: Vec<User>,
-    pub count: i32,
+    pub count: i64,
     pub rc: i32,
     pub message: String,
     pub pagination: Option<Pagination>,
@@ -13,7 +13,7 @@ pub struct UserResponse {
 }
 
 impl UserResponse {
-    pub fn new(results: Vec<User>, count: i32, pagination: Option<Pagination>, metadata: Option<ResponseMetadata>) -> Self {
+    pub fn new(results: Vec<User>, count: i64, pagination: Option<Pagination>, metadata: Option<ResponseMetadata>) -> Self {
         Self { results, count, rc: 0, message: "OK".to_string(), pagination, metadata }
     }
